API Consumption, UI and SSG :rocket:
This commit implements the usage of json-server to use the server.json as an API for the podcasts.

It also implements the app's header and sidebar UI, although the logic and interactions is still not implemented.

The app's homepage also has been implemented with Static Site Generation, to improve performance.