Walking the extra mile  :running:
Implementing "Repeat all" and improving Shuffle

To add my touch to the project, I chose to focus on small usability and user experience points.

In this commit, I added another repetition mode. When on regular repeat, the list will play through to the end, then play from the beginning. When on "Repeat one" mode, that podcast will be repeated.

The shuffle functionality has also been improved to resemble most media apps: the next episode doesn't get randomly assigned at the time of playing, but instead, the whole list is shuffled, which allows the user to skip forwards or backwards in the shuffled list.

Another improvement to the experience is the ability to go to the home page by clicking the website logo, a common practice in lots of websites. This was implemented in an earlier commit.